+++ backend/app/services/business/__init__.py
@@ -5,6 +5,25 @@ from .material_service import MaterialService
 from .instrument_service import InstrumentService
 from .qc_method_service import MethodParameterService, MethodService, MethodVersionService, TestService
 from .specification_service import SpecificationLimitService, SpecificationService, SpecificationTestService, SpecificationVersionService
-from .sample_service import SampleAPIService, SampleService, SampleTestService
+from .sample_service import SampleService, SampleAPIService, SampleTestService
+from .sample_test_assignment_service import SampleTestAssignmentService
 
-__all__ = ["InstrumentTypeService", "LocationService", "ManufacturerService", "MaterialService", "InstrumentService", "TestService", "MethodService", "MethodVersionService", "MethodParameterService", "SpecificationService", "SpecificationVersionService", "SpecificationTestService", "SpecificationLimitService", "SampleService", "SampleTestService", "SampleAPIService"]
+__all__ = [
+    "InstrumentTypeService",
+    "LocationService",
+    "ManufacturerService",
+    "MaterialService",
+    "InstrumentService",
+    "TestService",
+    "MethodService",
+    "MethodVersionService",
+    "MethodParameterService",
+    "SpecificationService",
+    "SpecificationVersionService",
+    "SpecificationTestService",
+    "SpecificationLimitService",
+    "SampleService",
+    "SampleAPIService",
+    "SampleTestService",
+    "SampleTestAssignmentService",
+]